- added redis for shop
- added redis LimitedProduct management
- added length check
- modified LimitedProductManageService findall
- added shop of limited product
- moved flash sale to new modules
- implemented flash sale module
- modified user wallet type, from Long to Float
- Added flash sale frontend
- added userId to login return content
- fixed typo
- added delay queue as message queue
- added Snowflake Algo as ID generator
- modified flash sale related beans
- modified flash sale modules
- modified ScheduledThreadPool size to 50
- modified flash sale module to become complete
- modified flash sale frontend to become complete
- added product status display
- connected azure redis
- implemented limited product list and add
- Merge remote-tracking branch 'origin/main' into gch

+++ backend/common-api/src/main/java/com/fellaverse/backend/dto/UserBasicInfoDTO.java
@@ -29,7 +29,7 @@ public class UserBasicInfoDTO implements Serializable {
     @NotBlank(groups = ValidGroup.Crud.Create.class, message = "Phone number cannot be blank")
     private String phoneNumber;
     @NotNull(message = "ID cannot be null")
-    private Long wallet;
+    private Float wallet;
     @NotNull(groups = ValidGroup.Crud.Create.class, message = "status annot be null")
     @EnumString(value = {"normal","locked","unknown"}, groups = ValidGroup.Crud.Update.class, message="invalid status")
     private UserStatus status;
